backend/requirements.txt
fastapi
uvicorn
sqlmodel
sqlalchemy
pydantic
pandas
openpyxl
httpx
python-dotenv
